+++ extension/src/setup/setupmgr.ts
@@ -12,6 +12,10 @@
  * limitations under the License.
  */
 
+import * as path from 'path' ;
+import * as fs from 'fs' ;
+import * as os from 'os' ;
+
 import { MTBAssistObject } from "../extobj/mtbassistobj";
 import { MtbManagerBase } from "../mgrbase/mgrbase";
 import { IDCLauncher } from "./launcher";
@@ -23,9 +27,7 @@ import { SetupProgram } from "../comms";
 import { ModusToolboxEnvironment } from "../mtbenv";
 import { MTBRunCommandOptions } from "../mtbenv/mtbenv/mtbenv";
 import { InstalledFeature } from "./installedfeature";
-import * as path from 'path' ;
-import * as fs from 'fs' ;
-import * as os from 'os' ;
+
 
 //
 // AppData/Local/Infineon_Technologies_AG/Infineon-Toolbox/Tools/ ...
@@ -56,6 +58,10 @@ export class SetupMgr extends MtbManagerBase {
     private static readonly mtbModusToolbox = 'com.ifx.tb.tool.modustoolbox' ;
     private static readonly downloadRatio = 0.8;
     private static readonly localIDCServiceRequestTimeout = 30000; // 30 seconds
+
+    private static readonly win32IDCDownloadURL = 'https://itools.infineon.com/itblauncher/service/itb-launcher-service-setup.exe' ;
+    private static readonly darwinIDCDownloadURL = 'https://itools.infineon.com/itblauncher/service/itb-launcher-service-setup.pkg' ;
+    private static readonly linuxIDCDownloadURL = 'https://itools.infineon.com/itblauncher/service/itb-launcher-service-setup.deb' ;
 
     private static readonly requiredFeatures : string[] = [
         'com.ifx.tb.tool.modustoolboxedgeprotectsecuritysuite',
@@ -223,22 +229,6 @@ export class SetupMgr extends MtbManagerBase {
         return basea.localeCompare(baseb);
     }
 
-    private findToolsDirFromPath(p: string) : string | undefined {
-        while (true) {
-            let bname = path.basename(p) ;
-            if (/^tools_[0-9]+.[0-9]+$/.test(bname)) {
-                return p ;
-            }
-            let parent = path.dirname(p) ;
-            if (parent === p) {
-                break ;
-            }
-            p = parent ;
-        }
-
-        return undefined ;
-    }
-
     public async initializeLocal() : Promise<void> {
         let ret = new Promise<void>((resolve, reject) => {
             this.registry_.initialize()
@@ -318,6 +308,19 @@ export class SetupMgr extends MtbManagerBase {
         return ret ;
     }
 
+    public findToolById(id: string) : string | undefined {
+        let ret : string | undefined ;
+        if (this.installed_.has(id)) {
+            let arr = this.installed_.get(id) ;
+            if (arr && arr.length > 0) {
+                let myarray = [...arr] ;
+                myarray.sort((a, b) => MTBVersion.compare(b.version, a.version)) ;
+                return myarray[0].installPath ;
+            }
+        }
+        return ret ;
+    }
+
     public findNeededTools() : SetupProgram[] {
         return [...this.checkNeededTools(SetupMgr.requiredFeatures, true)] ;
     }
@@ -348,6 +351,134 @@ export class SetupMgr extends MtbManagerBase {
             });
         });
 
+        return ret ;
+    }
+
+    public installIDCService() : Promise<void> {
+        if (process.platform === 'win32') {
+            return this.installIDCServiceWin32();
+        } else if (process.platform === 'darwin') {
+            return this.installIDCServiceDarwin();
+        } else if (process.platform === 'linux') {
+            return this.installIDCServiceLinux();
+        } else {
+            return Promise.reject(new Error(`Unsupported platform: ${process.platform}`));
+        }
+    }
+
+    private downloadFile(url: string, dest: string) : Promise<void> {
+        let ret = new Promise<void>((resolve, reject) => {
+            const file = fs.createWriteStream(dest);    
+            fetch(url)
+            .then((res: Response) => {
+                if (!res.ok) {
+                    reject(new Error(`Failed to download file: ${res.status} ${res.statusText}`));
+                    return;
+                }
+                res.body!.pipe(file);
+                res.body!.on('error', (err) => {
+                    reject(err);
+                });
+                file.on('finish', () => {
+                    file.close();
+                    resolve();
+                });
+                file.on('error', (err) => {
+                    fs.unlink(dest, () =>
+                        reject(err)
+                    );
+                });
+            })
+            .catch((err) => {
+                reject(err);
+            });
+        });
+        return ret ;
+    }
+
+    private installIDCServiceWin32() : Promise<void> {
+        let ret = new Promise<void>((resolve, reject) => {
+            let dest = path.join(os.tmpdir(), 'itb-launcher-service.exe') ;
+            this.emit('addStatusLine', 'Downloading IDC service...') ;
+            this.downloadFile(SetupMgr.win32IDCDownloadURL, dest)
+            .then(() => {
+                // Now install the service
+                let args: string[] = [] ;
+                args.push('/SP-') ;
+                args.push('/VERYSILENT') ;
+                args.push('/SUPPRESSMSGBOXES') ;
+                args.push('/CURRENTUSER') ;
+                this.emit('addStatusLine', 'Installing IDC service...') ;
+                ModusToolboxEnvironment.runCmdCaptureOutput(this.logger, dest, args, {})
+                .then((result) => { 
+                    if (result && result[0] === 0) {
+                        resolve() ;
+                    } else {
+                        reject(new Error('Failed to install IDC service')) ;
+                    }
+                })
+                .catch((err) => { 
+                    this.emit('addStatusLine', 'Error installing IDC service') ;
+                    reject(err) ;
+                }) ;
+            })
+            .catch((err) => { 
+                this.emit('addStatusLine', 'Error downloading IDC service') ;
+                reject(err) ;
+            }) ;
+        }) ;
+        return ret ;
+    }
+
+    private installIDCServiceDarwin() : Promise<void> {
+        let ret = new Promise<void>((resolve, reject) => {
+            this.emit('addStatusLine', 'Downloading IDC service...') ;
+            let target = path.join(os.tmpdir(), 'itb-launcher-service.pkg') ;            
+            this.downloadFile(SetupMgr.darwinIDCDownloadURL, target)
+            .then(async () => {
+                let passwd = await this.ext.getPasswordFromUser() ;
+                if (passwd === undefined) {
+                    reject(new Error('Password is install IDC service.'));
+                }
+
+                this.emit('addStatusLine', 'Installing IDC service...') ;
+                this.runInstallerDarwin(target, 'itb-launcher-service', '1.0', passwd)
+                .then(() => {
+                    resolve();
+                })
+                .catch((err) => {
+                    reject(err);
+                });
+            })
+            .catch((err) => {
+                reject(err);
+            });
+        });
+        return ret;
+    }
+
+    private installIDCServiceLinux() : Promise<void> {
+        let ret = new Promise<void>((resolve, reject) => {
+            this.emit('addStatusLine', 'Downloading IDC service...') ;
+            let target = path.join(os.tmpdir(), 'itb-launcher-service.deb') ;
+            this.downloadFile(SetupMgr.linuxIDCDownloadURL, target)
+            .then(async () => {
+                let password = await this.ext.getPasswordFromUser() ;
+                if (password === undefined) {
+                    reject(new Error('Password is install IDC service.'));
+                }
+                let args = ['dpkg', '-i', target] ;
+                let sudopwd = [ password! ] ;
+                let cmd = '/usr/bin/sudo' ;          
+                let opts: MTBRunCommandOptions = {
+                    stdin: sudopwd,
+                };                                      
+                ModusToolboxEnvironment.runCmdCaptureOutput(this.logger, cmd, args , opts);
+            })
+            .catch((err) => {
+                reject(err);
+            });
+        }) ;
         return ret ;
     }
 
@@ -644,11 +775,25 @@ export class SetupMgr extends MtbManagerBase {
                 return ;
             }
 
+            this.runInstallerDarwin(p, id, version, password)
+            .then(() => {
+                resolve();
+            })
+            .catch((err) => {
+                reject(err);
+            });
+
+        });
+        return ret;
+    }
+
+    private runInstallerDarwin(p: string, id: string, version: string, password?: string) : Promise<void> {
+        let ret = new Promise<void>((resolve, reject) => {
             let cmd ;
             let sudopwd: string[] | undefined ;
             let args: string[] ;
 
-            if (this.mtbLocation_?.startsWith('/Applications')) {
+            if (password) {
                 args = ['-S', '/usr/sbin/installer', '-target', '/Applications', '-pkg', p] ;                
                 sudopwd = [ password! ] ;
                 cmd = '/usr/bin/sudo' ;
@@ -678,9 +823,9 @@ export class SetupMgr extends MtbManagerBase {
             })
             .catch((err) => {
                 reject(new Error(`Failed to install feature ${id} - ${version} - ${err.message}`));
-            });
-        });
-        return ret;
+            });            
+        }) ;
+        return ret ;
     }
 
     private isPasswordError(text: string[]) : boolean {